+++ processes/gutenberg.py
@@ -19,8 +19,11 @@ class GutenbergProcess(CoreProcess):
         'pgterms': 'http://www.gutenberg.org/2009/pgterms/'
     }
 
-    def __init__(self, process, customFile, ingestPeriod):
-        super(GutenbergProcess, self).__init__(process, customFile, ingestPeriod)
+    def __init__(self, *args):
+        super(GutenbergProcess, self).__init__(*args[:3])
+
+        self.ingestOffset = int(args[4]) or 0
+        self.ingestLimit = (int(args[3]) + self.ingestOffset) or 5000
 
         # Connect to database
         self.generateEngine()
@@ -57,17 +60,23 @@ class GutenbergProcess(CoreProcess):
             orderDirection = 'ASC'
             orderField = 'CREATED_AT'
 
-        manager = GutenbergManager(orderDirection, orderField, startTimestamp, 100)
+        pageSize = 100
+        currentPosition = 0
+        manager = GutenbergManager(orderDirection, orderField, startTimestamp, pageSize)
 
         while True:
             continuation = manager.fetchGithubRepoBatch()
+
+            currentPosition += pageSize
+            if currentPosition <= self.ingestOffset: continue
+
             manager.fetchMetadataFilesForBatch()
 
             self.processGutenbergBatch(manager.dataFiles)
 
             manager.resetBatch()
 
-            if not continuation: break
+            if not continuation or currentPosition >= self.ingestLimit: break
 
     def processGutenbergBatch(self, dataFiles):
         for (gutenbergRDF, gutenbergYAML) in dataFiles:
@@ -86,20 +95,25 @@ class GutenbergProcess(CoreProcess):
 
     def storeEpubsInS3(self, gutenbergRec):
         for i, epubItem in enumerate(gutenbergRec.record.has_part):
-            pos, gutenbergURL, source, mediaType, flags = epubItem.split('|')
+            pos, gutenbergURL, source, mediaType, flagStr = epubItem.split('|')
 
             epubIDParts = re.search(r'\/([0-9]+).epub.([a-z]+)$', gutenbergURL)
             gutenbergID = epubIDParts.group(1)
             gutenbergType = epubIDParts.group(2)
 
-            bucketLocation = 'epubs/{}/{}_{}.epub'.format(source, gutenbergID, gutenbergType)
+            flags = json.loads(flagStr)
 
-            s3URL = 'https://{}.s3.amazonaws.com/{}'.format(
-                self.s3Bucket, bucketLocation
-            )
-            gutenbergRec.record.has_part[i] = '|'.join([pos, s3URL, source, mediaType, flags])
+            if flags['download'] is True:
+                bucketLocation = 'epubs/{}/{}_{}.epub'.format(source, gutenbergID, gutenbergType)
+            else:
+                bucketLocation = 'epubs/{}/{}_{}/oebps/content.opf'.format(source, gutenbergID, gutenbergType)
 
-            self.sendFileToProcessingQueue(gutenbergURL, bucketLocation)
+            s3URL = 'https://{}.s3.amazonaws.com/{}'.format(self.s3Bucket, bucketLocation)
+
+            gutenbergRec.record.has_part[i] = '|'.join([pos, s3URL, source, mediaType, flagStr])
+
+            if flags['download'] is True:
+                self.sendFileToProcessingQueue(gutenbergURL, bucketLocation)
 
     def addCoverAndStoreInS3(self, gutenbergRec, yamlData):
         for coverData in yamlData['covers']: