+++ processes/s3Files.py
@@ -9,8 +9,8 @@ from managers import S3Manager, RabbitMQManager
 
 
 class S3Process(CoreProcess):
-    def __init__(self, process, customFile, ingestPeriod):
-        super(S3Process, self).__init__(process, customFile, ingestPeriod)
+    def __init__(self, *args):
+        super(S3Process, self).__init__(*args[:3])
 
     def runProcess(self):
         self.receiveAndProcessMessages()