+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## Unreleased -- v0.0.5
 ### Added
 - Added multiprocessing to S3FileProcess
+- `limit` and `offset` as optional arguments to the main process
+- "Exploded" versions of ePub files in S3
 ### Fixed
 - Add ability to define default role for agents in ElasticSearch manager
 - Clean up dates as they are processed to remove duplicate values
 - Fix regression in API with ElasticSearch aggregation
+- Improved error handling in HathiTrust processor
 
 ## 2020-12-28 -- v0.0.4
 ### Added